update Primer3 parser
The Primer3 was upgraded to version 2.0, which introduced some new keys in the output. These are now handled.
The error messages was improved to make more sense in future incidents
When an error occurs the program will now output a pkl file with the input from Primer3. This will help later debugging/testing.
The function also undergoes switch from 3 space indentation to 4